# docs/swagger.yaml
basePath: /api
definitions:
...
  models.CreateRentalRequest:
    properties:
      comment:
        description: Статус аренды
        type: string
      end_date:
        description: Дата завершения аренды
        type: string
      field_id:
        description: Идентификатор площадки
        type: integer
      id:
        description: Идентификатор
        type: integer
      start_date:
        description: Дата начала аренды
        type: string
      team_id:
        description: Идентификатор команды
        type: integer
    required:
    - end_date
    - field_id
    - start_date
    - team_id
    type: object
...
  models.ErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/models.ValidationErrorResponse'
        type: array
      message:
        type: string
      statusCode:
        type: integer
    type: object
  models.FieldView:
    properties:
      address:
        type: string
      city:
        type: string
      created_at:
        type: string
      description:
        type: string
      display:
        type: boolean
      dressing:
        type: boolean
      for_disabled:
        type: boolean
      id:
        type: integer
      info:
        type: string
      location:
        type: string
      logo:
        allOf:
        - $ref: '#/definitions/models.Media'
        description: Логотип
      media:
        description: Медиа
        items:
          $ref: '#/definitions/models.Media'
        type: array
      name:
        type: string
      parking:
        type: boolean
      places:
        type: integer
      slug:
        type: string
      square:
        type: integer
      status:
        type: integer
      toilet:
        type: boolean
    type: object
...
  models.Media:
    properties:
      created_at:
        type: string
      ext:
        type: string
      id:
        type: integer
      name:
        type: string
      path:
        type: string
      size:
        type: integer
    type: object
  models.RentalView:
    properties:
      comment:
        description: Комментарий
        type: string
      created_at:
        description: Дата создания
        type: string
      duration:
        description: Длительность аренды (например, в часах)
        type: integer
      end_date:
        description: Дата завершения аренды
        type: string
      field:
        allOf:
        - $ref: '#/definitions/models.FieldView'
        description: Пощадка
      id:
        description: Идентификатор
        type: integer
      start_date:
        description: Дата начала аренды
        type: string
      status:
        description: Статус аренды
        type: integer
      team:
        allOf:
        - $ref: '#/definitions/models.TeamView'
        description: Команда
      user:
        allOf:
        - $ref: '#/definitions/models.UserView'
        description: Пользователь
    type: object
  models.Role:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.TeamView:
    properties:
      city:
        description: Город
        type: string
      created_at:
        description: Дата создания
        type: string
      description:
        description: Описание
        type: string
      disability_category:
        description: Категория инвалидности
        type: string
      id:
        type: integer
      logo:
        allOf:
        - $ref: '#/definitions/models.Media'
        description: Логотип
      media:
        description: Медиа
        items:
          $ref: '#/definitions/models.Media'
        type: array
      name:
        description: Название
        type: string
      participant_count:
        description: Кол-во участников
        type: integer
      responsible:
        allOf:
        - $ref: '#/definitions/models.UserView'
        description: Ответственный
      status:
        description: Статус
        type: integer
      uniform_color:
        description: Цвет формы
        type: string
    type: object
...
  models.UserView:
    properties:
      city:
        description: Город
        type: string
      created_at:
        description: Дата создания
        type: string
      email:
        description: Email
        type: string
      id:
        type: integer
      logo:
        description: Логотип
        type: string
      media:
        description: Медиа
        type: string
      name:
        description: ФИО
        type: string
      phone:
        description: Телефон
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        description: Роль
      status:
        description: Статус
        type: integer
    type: object
  models.ValidationErrorResponse:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Приложения бронирования спортивных площадок.
  title: Спортивный город
  version: "1.0"
paths:
...
  /api/media/preloader:
    post:
      description: Загрузка медиафайла
      parameters:
      - description: Загруженный файл
        in: formData
        name: file
        required: true
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Media'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "415":
          description: Unsupported Media Type
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Загрузить медиафайл
      tags:
      - Медиафайлы
  /api/rentals:
    get:
      consumes:
      - application/json
      description: Получение списка всех аренд
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RentalView'
            type: array
        "400":
          description: Bad Request
          schema:
            type: Bad
        "500":
          description: Internal Server Error
          schema:
            type: Internal
      summary: Возвращает список всех аренд
      tags:
      - Аренда
    post:
      description: Создание новой аренды
      parameters:
      - description: Данные для создания новой аренды
        in: body
        name: createRental
        required: true
        schema:
          $ref: '#/definitions/models.CreateRentalRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.RentalView'
        "422":
          description: Unprocessable Entity
          schema:
            type: Unprocessable
      summary: Создание новой аренды
      tags:
      - Аренда
  /api/rentals/{id}:
    delete:
      description: Удаление аренды по идентификатору
      parameters:
      - description: ID аренды
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: "No"
        "404":
          description: Not Found
          schema:
            type: Not
      summary: Удаляет аренду по ID
      tags:
      - Аренда
    get:
      description: Получение информации об аренде по идентификатору
      parameters:
      - description: ID аренды
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RentalView'
        "400":
          description: Bad Request
          schema:
            type: Bad
        "404":
          description: Not Found
          schema:
            type: Not
      summary: Возвращает информацию об аренде по ID
      tags:
      - Аренда
...
swagger: "2.0"